Replace AutoCompleteView with custom filter on  ContactListFragment
